[advanced] 타입과 더미데이터 분리
