+++ packages/assignment-7/src/App.tsx
@@ -40,41 +40,14 @@ import {
   DeleteIcon,
   EditIcon,
 } from '@chakra-ui/icons';
-
-type RepeatType = 'none' | 'daily' | 'weekly' | 'monthly' | 'yearly';
-
-interface RepeatInfo {
-  type: RepeatType;
-  interval: number;
-  endDate?: string;
-}
-
-export interface Event {
-  id: number;
-  title: string;
-  date: string;
-  startTime: string;
-  endTime: string;
-  description: string;
-  location: string;
-  category: string;
-  repeat: RepeatInfo;
-  notificationTime: number; // 분 단위로 저장
-}
-
-const categories = ['업무', '개인', '가족', '기타'];
-
-const weekDays = ['일', '월', '화', '수', '목', '금', '토'];
-
-const notificationOptions = [
-  { value: 1, label: '1분 전' },
-  { value: 10, label: '10분 전' },
-  { value: 60, label: '1시간 전' },
-  { value: 120, label: '2시간 전' },
-  { value: 1440, label: '1일 전' },
-];
-
-const dummyEvents: Event[] = [];
+import {
+  categories,
+  dummyEvents,
+  Event,
+  notificationOptions,
+  RepeatType,
+  weekDays,
+} from './hooks/useEvents';
 
 const fetchHolidays = async (year: number, month: number) => {
   try {
